+++ CHANGES.rst
@@ -27,6 +27,7 @@ Bugfixes & Improvements
 - Prevent auto creation of invalid links in comments (`#995`_, pkvach)
 - Fix W3C Validation issues (`#999`_, pkvach)
 - Handle deleted comments in Disqus migration (`#994`_, pkvach)
+- Fix total comments count calculation (`#997`_, pkvach)
 
 .. _#951: https://github.com/posativ/isso/pull/951
 .. _#967: https://github.com/posativ/isso/pull/967
@@ -34,6 +35,7 @@ Bugfixes & Improvements
 .. _#995: https://github.com/isso-comments/isso/pull/995
 .. _#999: https://github.com/isso-comments/isso/pull/999
 .. _#994: https://github.com/isso-comments/isso/pull/994
+.. _#997: https://github.com/isso-comments/isso/pull/997
 
 0.13.1.dev0 (2023-02-05)
 ------------------------